Add multiple optimizations

+++ src/tmp/Ant.java
@@ -19,8 +19,9 @@ public class Ant {
         this.totalConstraintsNumber = ant.totalConstraintsNumber;
     }
 
-    public int getDistance() {
+    public double getDistance() {
         return this.failedRelationNumber + this.failedConstraintsNumber;
+//        return 0.45F * (float)this.failedConstraintsNumber / (float)this.totalConstraintsNumber + 0.55F * (float)this.failedRelationNumber;
     }
 
     public Ant() {
